+++ pkg/providers/reportinggroups/testdata/non_default_edgerc/variables.tf
@@ -1,0 +1,9 @@
+variable "edgerc_path" {
+  type    = string
+  default = "/non/default/path/.edgerc"
+}
+
+variable "config_section" {
+  type    = string
+  default = "other-section"
+}